Feature/create entity

+++ src/Dal/DalTest.java
@@ -31,9 +31,9 @@ public class DalTest {
                 test.setIdTest(rs.getInt("idTest"));
                 test.setLibelle(rs.getString("libelle"));
                 test.setDescription(rs.getString("description"));
-                test.setDuree(rs.getDouble("duree"));
-                test.setSeuilHaut(rs.getDouble("seuil_haut"));
-                test.setSeuilBas(rs.getDouble("seuil_bas"));
+                test.setDuree(rs.getInt("duree"));
+                test.setSeuilHaut(rs.getInt("seuil_haut"));
+                test.setSeuilBas(rs.getInt("seuil_bas"));
 
                 listeTest.add(test);
             }
